Convert app clean-up step into fixture
