+++ tests/test_main.py
@@ -9,8 +9,8 @@ import tarfile
 from pathlib import Path
 
 import pytest
-from conftest import CONDA_EXE, _get_shortcut_dirs, menuinst_pkg_specs, run_conda
 from ruamel.yaml import YAML
+from utils import CONDA_EXE, run_conda
 
 HERE = Path(__file__).parent
 
@@ -169,101 +169,66 @@ def test_extract_conda_pkgs_num_processors(tmp_path: Path):
     )
 
 
-_pkg_specs_params = pytest.mark.parametrize(
-    "pkg_spec, shortcut_path", menuinst_pkg_specs
-)
-
-
-@_pkg_specs_params
-def test_menuinst_conda(tmp_path: Path, pkg_spec: str, shortcut_path: dict[str, str]):
+def test_menuinst_conda(tmp_path: Path, clean_shortcuts: dict[str, list[Path]]):
     "Check 'regular' conda can process menuinst JSONs"
-
-    # If the test app already exists, this test will fail,
-    # so clean up before and after the run.
-    def _clean_macos_apps(shortcuts: list[Path]):
-        if not sys.platform == "darwin":
-            return
-        for shortcut in shortcuts:
-            if shortcut.exists():
-                shutil.rmtree(shortcut)
 
     env = os.environ.copy()
     env["CONDA_ROOT_PREFIX"] = sys.prefix
-    # The shortcut will take 'root_prefix' as the base, but conda-standalone
-    # sets that to its temporary 'sys.prefix' as provided by the pyinstaller
-    # self-extraction. We override it via 'CONDA_ROOT_PREFIX' in the same
-    # way 'constructor' will do it.
-    variables = {"base": Path(sys.prefix).name, "name": tmp_path.name}
-    shortcuts = [
-        folder / shortcut_path[sys.platform].format(**variables)
-        for folder in _get_shortcut_dirs()
-    ]
-    _clean_macos_apps(shortcuts)
-    try:
-        process = run_conda(
-            "create",
-            "-vvv",
-            "-p",
-            tmp_path,
-            "-y",
-            pkg_spec,
-            "--no-deps",
-            env=env,
-            capture_output=True,
-            text=True,
-            check=True,
-        )
-        print(process.stdout)
-        print(process.stderr, file=sys.stderr)
-        assert "menuinst Exception" not in process.stdout
-        assert list(tmp_path.glob("Menu/*.json"))
-        assert any(shortcut.exists() for shortcut in shortcuts)
-        process = run_conda(
-            "remove",
-            "-vvv",
-            "-p",
-            tmp_path,
-            "-y",
-            pkg_spec.split("::")[-1],
-            env=env,
-            capture_output=True,
-            text=True,
-            check=True,
-        )
-        print(process.stdout)
-        print(process.stderr, file=sys.stderr)
-        assert not any(shortcut.exists() for shortcut in shortcuts)
-    finally:
-        _clean_macos_apps(shortcuts)
-
-
-@_pkg_specs_params
-def test_menuinst_constructor(tmp_path: Path, pkg_spec: str, shortcut_path: str):
-    "The constructor helper should also be able to process menuinst JSONs"
-
-    # If the test app already exists, this test will fail,
-    # so clean up before and after the run.
-    def _clean_macos_apps(shortcuts: list[Path]):
-        if not sys.platform == "darwin":
-            return
-        for shortcut in shortcuts:
-            if shortcut.exists():
-                shutil.rmtree(shortcut)
-
-    run_kwargs = dict(capture_output=True, text=True, check=True)
-    variables = {"base": Path(sys.prefix).name, "name": tmp_path.name}
-    shortcuts = [
-        folder / shortcut_path[sys.platform].format(**variables)
-        for folder in _get_shortcut_dirs()
-    ]
-    _clean_macos_apps(shortcuts)
     process = run_conda(
         "create",
         "-vvv",
         "-p",
         tmp_path,
         "-y",
-        pkg_spec,
+        *clean_shortcuts.keys(),
+        "--no-deps",
+        env=env,
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    print(process.stdout)
+    print(process.stderr, file=sys.stderr)
+    assert "menuinst Exception" not in process.stdout
+    assert list(tmp_path.glob("Menu/*.json"))
+    shortcuts_found = [
+        package
+        for package, shortcuts in clean_shortcuts.items()
+        if any(shortcut.exists() for shortcut in shortcuts)
+    ]
+    assert sorted(shortcuts_found) == sorted(clean_shortcuts.keys())
+    process = run_conda(
+        "remove",
+        "-vvv",
+        "-p",
+        tmp_path,
+        "-y",
+        *[pkg_spec.split("::")[-1] for pkg_spec in clean_shortcuts.keys()],
+        env=env,
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+    print(process.stdout)
+    print(process.stderr, file=sys.stderr)
+    shortcuts_found = [
+        package
+        for package, shortcuts in clean_shortcuts.items()
+        if any(shortcut.exists() for shortcut in shortcuts)
+    ]
+    assert shortcuts_found == []
+
+
+def test_menuinst_constructor(tmp_path: Path, clean_shortcuts: dict[str, list[Path]]):
+    "The constructor helper should also be able to process menuinst JSONs"
+    run_kwargs = dict(capture_output=True, text=True, check=True)
+    process = run_conda(
+        "create",
+        "-vvv",
+        "-p",
+        tmp_path,
+        "-y",
+        *clean_shortcuts.keys(),
         "--no-deps",
         "--no-shortcuts",
         **run_kwargs,
@@ -274,40 +239,47 @@ def test_menuinst_constructor(tmp_path: Path, pkg_spec: str, shortcut_path: str)
 
     env = os.environ.copy()
     env["CONDA_ROOT_PREFIX"] = sys.prefix
-    try:
-        process = run_conda(
-            "constructor",
-            # Not supported in micromamba's interface yet
-            # use CONDA_ROOT_PREFIX instead
-            # "--root-prefix",
-            # sys.prefix,
-            "--prefix",
-            tmp_path,
-            "--make-menus",
-            **run_kwargs,
-            env=env,
-        )
-        print(process.stdout)
-        print(process.stderr, file=sys.stderr)
-        assert any(shortcut.exists() for shortcut in shortcuts)
+    process = run_conda(
+        "constructor",
+        # Not supported in micromamba's interface yet
+        # use CONDA_ROOT_PREFIX instead
+        # "--root-prefix",
+        # sys.prefix,
+        "--prefix",
+        tmp_path,
+        "--make-menus",
+        **run_kwargs,
+        env=env,
+    )
+    print(process.stdout)
+    print(process.stderr, file=sys.stderr)
+    shortcuts_found = [
+        package
+        for package, shortcuts in clean_shortcuts.items()
+        if any(shortcut.exists() for shortcut in shortcuts)
+    ]
+    assert sorted(shortcuts_found) == sorted(clean_shortcuts.keys())
 
-        process = run_conda(
-            "constructor",
-            # Not supported in micromamba's interface yet
-            # use CONDA_ROOT_PREFIX instead
-            # "--root-prefix",
-            # sys.prefix,
-            "--prefix",
-            tmp_path,
-            "--rm-menus",
-            **run_kwargs,
-            env=env,
-        )
-        print(process.stdout)
-        print(process.stderr, file=sys.stderr)
-        assert not any(shortcut.exists() for shortcut in shortcuts)
-    finally:
-        _clean_macos_apps(shortcuts)
+    process = run_conda(
+        "constructor",
+        # Not supported in micromamba's interface yet
+        # use CONDA_ROOT_PREFIX instead
+        # "--root-prefix",
+        # sys.prefix,
+        "--prefix",
+        tmp_path,
+        "--rm-menus",
+        **run_kwargs,
+        env=env,
+    )
+    print(process.stdout)
+    print(process.stderr, file=sys.stderr)
+    shortcuts_found = [
+        package
+        for package, shortcuts in clean_shortcuts.items()
+        if any(shortcut.exists() for shortcut in shortcuts)
+    ]
+    assert shortcuts_found == []
 
 
 def test_python():
